file structure changes in js
