+++ src/index.html
@@ -328,7 +328,7 @@
                   </button>
                 </form>
 
-                <div class="row">
+                <div class="row mb-4">
                   <div class="col-12 justify-content-between">
                     <button
                       type="button"
